requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
apscheduler==3.10.4
tensorflow==2.12.0
numpy==1.24.4
google-cloud-firestore==3.9.1
google-auth==2.24.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
